changed image locations
GitHub Pages did not allow for loading files from outside the `/doc` folder, so the links needed to be adapted. At the same time, it was better practice to keep a separation between the collection (with transcription and master files for images) and the website (with html pages and presentation copies in different sizes for the images).